+++ src/app/admin/usersTable/headers.ts
@@ -11,20 +11,3 @@ export const usersTableHeaderKeysAndLabels: TableHeaders<UserRow> = [
     ["createdAt", "Created At"],
     ["updatedAt", "Updated At"],
 ];
-
-export const formatTimestamp = (timestamp: number): string => {
-    if (isNaN(timestamp)) {
-        return "-";
-    }
-
-    return new Date(timestamp).toLocaleString("en-gb");
-};
-
-export const userTableColumnDisplayFunctions = {
-    createdAt: (createdAt: number | null) => {
-        return createdAt === null ? "-" : formatTimestamp(createdAt);
-    },
-    updatedAt: (updatedAt: number | null) => {
-        return updatedAt === null ? "-" : formatTimestamp(updatedAt);
-    },
-};
