+++ src/app/parcels/ActionBar/ActionAndStatusBar.tsx
@@ -3,11 +3,11 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 import Button from "@mui/material/Button";
+import { ParcelsTableRow } from "../parcelsTable/types";
 import Alert from "@mui/material/Alert";
 import Statuses, { StatusType, SaveParcelStatusResult } from "@/app/parcels/ActionBar/Statuses";
 import Actions from "@/app/parcels/ActionBar/Actions";
 import { ArrowDropDown } from "@mui/icons-material";
-import { ParcelsTableRow } from "../parcelsTable/types";
 
 export interface ActionAndStatusBarProps {
     fetchSelectedParcels: () => Promise<ParcelsTableRow[]>;
