+++ src/app/parcels/ActionBar/ActionModals/GeneralActionModal.tsx
@@ -1,11 +1,11 @@
 import styled from "styled-components";
+import { ParcelsTableRow } from "../../parcelsTable/types";
 import Modal from "@/components/Modal/Modal";
 import { ActionName } from "../Actions";
 import { StatusType } from "../Statuses";
 import React from "react";
 import { UpdateParcelStatuses } from "../ActionAndStatusBar";
 import { ErrorSecondaryText } from "@/app/errorStylingandMessages";
-import { ParcelsTableRow } from "../../parcelsTable/types";
 
 export interface ActionModalProps extends Omit<React.ComponentProps<typeof Modal>, "children"> {
     selectedParcels: ParcelsTableRow[];
