+++ src/app/parcels/parcelsTable/styles.ts
@@ -1,0 +1,38 @@
+import { styled } from "styled-components";
+
+export const PreTableControls = styled.div`
+    margin: 1rem;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+    align-items: stretch;
+    justify-content: space-between;
+`;
+
+export const parcelTableColumnStyleOptions = {
+    iconsColumn: {
+        width: "4rem",
+    },
+    fullName: {
+        minWidth: "8rem",
+    },
+    familyCategory: {
+        hide: 550,
+    },
+    addressPostcode: {
+        hide: 800,
+    },
+    deliveryCollection: {
+        minWidth: "6rem",
+    },
+    packingDate: {
+        hide: 800,
+    },
+    packingSlot: {
+        hide: 800,
+        minWidth: "6rem",
+    },
+    lastStatus: {
+        minWidth: "8rem",
+    },
+};
