+++ src/pdf/ShippingLabels/ShippingLabelsPdfButton.tsx
@@ -4,10 +4,10 @@ import React from "react";
 import supabase from "@/supabaseClient";
 import PdfButton from "@/components/PdfButton/PdfButton";
 import ShippingLabelsPdf, { ShippingLabelData } from "@/pdf/ShippingLabels/ShippingLabelsPdf";
+import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
 import { PdfDataFetchResponse } from "../common";
 import { Schema } from "@/databaseUtils";
 import { logErrorReturnLogId } from "@/logger/logger";
-import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
 
 const formatDatetime = (datetimeString: string | null): string => {
     if (datetimeString === null) {
